# web/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Shortlist",
    "version": "0.1.0b8"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/report": {
      "get": {
        "description": "The dashboard tracking report: headline counts for the chosen window with their change vs the\nprevious equal period, the landing rate over a matured cohort, watch momentum, per-user and\nper-row breakdowns, the titles landing best, requests that paid off, and a recent-watches feed.\n\nDeliberately a plain `def`, not `async def`: the computation is ~30 synchronous queries, one of\nthem a `GROUP BY user_id, tmdb_id, media_type` over the whole picks table. On the event loop that\nstalled SSE, `/api/system/health` and every other request for as long as it took; as a sync\nhandler Starlette runs it in a worker thread instead.",
        "operationId": "effectiveness_api_report_get",
        "parameters": [
          {
            "description": "Report window in days: 7, 30, 90, or 'all'.",
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "30",
              "description": "Report window in days: 7, 30, 90, or 'all'.",
              "title": "Window",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Effectiveness Api Report Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Effectiveness",
        "tags": [
          "report"
        ]
      }
    },
    "/api/report/deleted-rows": {
      "delete": {
        "description": "Permanently delete the pick history of rows that no longer exist. `slug` clears just one.\n\nThis is the one destructive action on the dashboard, so it is deliberately narrow:\n\n* Only slugs with no live Collection are eligible \u2014 the eligible set is recomputed here rather\n  than trusted from the request, so naming a live row's slug deletes nothing.\n* Only `picks` rows are touched. `deliveries` is left alone: it is the ledger of what still\n  exists on Plex, and clearing it would strand a real collection with nothing left to clean it up.\n* These picks disappear from everywhere they are counted \u2014 not just this dashboard. The per-user\n  lifetime stats and a person's pick history on their own page are read from the same rows, so the\n  UI has to say \"history\", not \"the numbers above\".",
        "operationId": "clear_deleted_rows_api_report_deleted_rows_delete",
        "parameters": [
          {
            "in": "query",
            "name": "slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Clear Deleted Rows Api Report Deleted Rows Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Deleted Rows",
        "tags": [
          "report"
        ]
      },
      "get": {
        "description": "Pick history belonging to rows that no longer exist, with how much of it there is.\n\nIts own endpoint rather than a field on the report: the report is windowed, and \"what can I clear\"\nis a question about ALL of a deleted row's history, not the last 30 days of it.",
        "operationId": "deleted_rows_api_report_deleted_rows_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Deleted Rows Api Report Deleted Rows Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Deleted Rows",
        "tags": [
          "report"
        ]
      }
    },
...
  }
}